fix bundle size issue
