+++ tina/collection/page.ts
@@ -2,7 +2,7 @@ import type { Collection } from "tinacms";
 import { heroBlockSchema } from "../../components/blocks/hero";
 import { contentBlockSchema } from "../../components/blocks/content";
 import { testimonialBlockSchema } from "../../components/blocks/testimonial";
-import { featureBlockSchema } from "../../components/blocks/features";
+import { featureBlockSchema } from "../../components/blocks/features.schema";
 
 const Page: Collection = {
   label: "Pages",
